Move the demo xblocks into their own package so that the xblock package can be installed without the demo xblocks.
