+++ demo_xblocks/README.rst
@@ -1,0 +1,22 @@
+Demo XBlocks
+============
+
+This package contains some sample XBlocks to illustrate the capabilities of
+the XBlock specification. The following blocks are included:
+
+  * **helloworld_demo:** A block which simply displays the text "Hello World"
+
+  * **html_demo:** A block which displays a specified block of HTML text
+
+  * **sequence_demo:** A container block which displays its children in a tabbed layout
+
+  * **vertical_demo:** A container block which displays its children in a vertical list
+
+  * **sidebar_demo:** A variant of the vertical block
+
+  * **problem_demo:** A block which presents a problem and grades student responses
+
+  * **slider_demo:** A block which presents a slider and remembers the student's setting
+
+  * **view_counter:** A block which shows the total number of times it has been viewed
+
