+++ demo_xblocks/demo_xblocks/public/images/correct-icon.png
