+++ demo_xblocks/demo_xblocks/slider.py
