+++ script/cibuild
@@ -29,4 +29,4 @@ diff-quality --violations=pylint --html-report diff_quality_pylint.html
 
 # Compute style violations
 pep8 > pep8.report || echo "Not pep8 clean"
-pylint -f parseable xblock workbench thumbs/thumbs/thumbs.py > pylint.report || echo "Not pylint clean"
+pylint -f parseable xblock demo-xblocks workbench thumbs/thumbs/thumbs.py > pylint.report || echo "Not pylint clean"
